entrypoints fixup, fix comment deletion buttons
