+++ app/frontend/entrypoints/events/show.ts
@@ -1,0 +1,1 @@
+import '@/comments.js';
