+++ app/views/events/_form.html.haml
@@ -1,4 +1,4 @@
-= vite_typescript_tag 'src/events'
+= vite_typescript_tag 'entrypoints/events/form'
 = vite_stylesheet_tag 'styles/events.scss'
 
 - # svg config for info circle
