+++ app/views/events/show.html.haml
@@ -1,4 +1,4 @@
-= vite_typescript_tag 'src/comments'
+= vite_typescript_tag 'entrypoints/events/show'
 = vite_stylesheet_tag 'styles/events.scss'
 - title @event.title
